ci-generator/requirements.txt
coverage==4.5.3
Jinja2==2.10.1
PyYAML==5.1
requests==2.22.0
requests-mock==1.6.0
